requirements.txt
numpy==1.22.3
transformers==4.19.2
pandas==1.4.2
flask==2.2.2
Werkzeug==2.3.7
requests==2.27.1
scipy==1.8.0
pymongo==4.1.1
sklearn==0.0
scikit-learn==1.1.0
nlpaug~=1.1.10
torch==1.11.0
kafka-python==2.0.2 
utils~=1.0.1